# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication/Authorization Service",
        "title": "Sentinel",
        "contact": {},
        "license": {
            "name": "AGPL-3.0 (With additional terms, see NOTICE file)",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth/oauth/google/callback": {
            "get": {
                "description": "Do not access this endpoint manually, google API will automatically redirect to it. E-Mail of google account MUST be verified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "third-party-auth",
                    "oauth"
                ],
                "summary": "Actual handler for auth via google account",
                "operationId": "google-login-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User credentials and audience",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
        "/auth/oauth/google/login": {
            "get": {
                "description": "Alternative login endpoint, redirects to the google OAuth2.0 endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "third-party-auth",
                    "oauth"
                ],
                "summary": "Login/Signup into the service using google account",
                "operationId": "google-login",
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responsebody.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error"
                },
                "message": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "responsebody.Token": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "expiresIn": {
                    "type": "integer",
                    "example": 600
                },
                "message": {
                    "type": "string",
                    "example": "hello"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token format: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRF_Cookie": {
            "description": "CSRF token in cookie. Required for state-changing requests, must match with token in X-CSRF-Token header",
            "type": "apiKey",
            "name": "_csrf",
            "in": "cookie"
        },
        "CSRF_Header": {
            "description": "CSRF token in request header. Required for state-changing requests, must match with token in cookie",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}